Refactor conversation message sync naming
