+++ src/features/conversations/components/MessageList.tsx
@@ -20,7 +20,7 @@ import {
 import type {
   ChatMessage,
   MessageAttachment,
-} from '@stores/conversation/interfaces.js';
+} from '@stores/conversation/types.js';
 import type { ConversationId } from '@stores/conversation/types.js';
 import { createConversationFileObjectUrl } from '@stores/files-store';
